Fix and simplify Windows example configs
Some of the connection settings were in the children section.

However, configuring reauth_time is useless with EAP authentication
and clients that don't support RFC 4478 (such as Windows, no idea
if it could even handle server-initiated reauth with machine
certs), and rekey_time's default value is already one hour.

Finally, explicit proposals are not necessary to accept the Windows
defaults and they prevent stronger proposals if the configs are used
for non-Windows clients or PowerShell is used to enable stronger
algorithms on Windows clients.  So remove the proposals but add
a NOTE to inform the users. strongswan/strongswan-docs@0ab3ef5f1b1cba559860bd478cba1947902106b3